# docs/architecture/control-plane/operator-ui-mvp-openapi.yaml
openapi: 3.1.0
info:
  title: OneFlow Control Plane - Monitoring First API
  version: 0.1.0
  description: |
    Current contract for the monitoring-first control-plane starter implemented by
    com.etl.controlplane.ControlPlaneApiApplication.
servers:
  - url: /api/v1
    description: Relative base path for the monitoring-first control-plane API
...
paths:
...
  /schedules:
    get:
      tags: [Schedules]
      operationId: listSchedules
      summary: List persisted schedules
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 25
      responses:
        '200':
          description: Schedule list response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleListResponse'
    post:
      tags: [Schedules]
      operationId: createSchedule
      summary: Create one persisted schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
      responses:
        '201':
          description: Created schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleViewResponse'
        '400':
          description: Invalid request
        '409':
          description: Duplicate schedule key
  /schedules/{scheduleId}:
    get:
      tags: [Schedules]
      operationId: getScheduleById
      summary: Get one schedule by id
      parameters:
        - $ref: '#/components/parameters/scheduleId'
      responses:
        '200':
          description: Schedule detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleViewResponse'
        '404':
          description: Schedule not found
    put:
      tags: [Schedules]
      operationId: updateSchedule
      summary: Update one schedule
      parameters:
        - $ref: '#/components/parameters/scheduleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScheduleRequest'
      responses:
        '200':
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleViewResponse'
        '400':
          description: Invalid request
        '404':
          description: Schedule not found
  /schedules/{scheduleId}:enable:
    post:
      tags: [Schedules]
      operationId: enableSchedule
      summary: Enable one schedule
      parameters:
        - $ref: '#/components/parameters/scheduleId'
      responses:
        '200':
          description: Schedule state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleStateChangeResponse'
        '404':
          description: Schedule not found
  /schedules/{scheduleId}:disable:
    post:
      tags: [Schedules]
      operationId: disableSchedule
      summary: Disable one schedule
      parameters:
        - $ref: '#/components/parameters/scheduleId'
      responses:
        '200':
          description: Schedule state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleStateChangeResponse'
        '404':
          description: Schedule not found
  /schedules/{scheduleId}:pause:
    post:
      tags: [Schedules]
      operationId: pauseSchedule
      summary: Pause one schedule
      parameters:
        - $ref: '#/components/parameters/scheduleId'
      responses:
        '200':
          description: Schedule state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleStateChangeResponse'
        '404':
          description: Schedule not found
  /schedules/{scheduleId}:resume:
    post:
      tags: [Schedules]
      operationId: resumeSchedule
      summary: Resume one schedule
      parameters:
        - $ref: '#/components/parameters/scheduleId'
      responses:
        '200':
          description: Schedule state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleStateChangeResponse'
        '404':
          description: Schedule not found
  /schedules/{scheduleId}/trigger-events:
    get:
      tags: [Schedules]
      operationId: listScheduleTriggerEvents
      summary: List recent trigger events for one schedule id
      parameters:
        - $ref: '#/components/parameters/scheduleId'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - in: query
          name: size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
          description: Optional alias that takes precedence over `limit`.
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Zero-based page number.
      responses:
        '200':
          description: Trigger event history entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerEventListResponse'
        '404':
          description: Schedule not found
...
components:
  parameters:
...
    scheduleId:
      in: path
      name: scheduleId
      required: true
      schema:
        type: string
  schemas:
...
    TriggerEventView:
      type: object
      properties:
        triggerEventId:
          type: string
        jobKey:
          type: string
        decisionStatus:
          type: string
        reason:
          type: string
        requestedBy:
          type: string
        requestedAt:
          type: string
          format: date-time
        launchedRunId:
          oneOf:
            - type: string
            - type: 'null'
        message:
          type: string
      required: [triggerEventId, jobKey, decisionStatus, reason, requestedBy, requestedAt, launchedRunId, message]
    TriggerEventListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TriggerEventView'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
      required: [items, page, size, totalItems]
    ScheduleViewResponse:
      type: object
      properties:
        scheduleId:
          type: string
        scheduleKey:
          type: string
        selectedJobKey:
          type: string
        expression:
          type: string
        timezone:
          type: string
        enabled:
          type: boolean
        paused:
          type: boolean
        description:
          type: string
        updatedAt:
          type: string
          format: date-time
        lastAcceptedDueAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        nextDueAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
      required: [scheduleId, scheduleKey, selectedJobKey, expression, timezone, enabled, paused, description, updatedAt, lastAcceptedDueAt, nextDueAt]
    ScheduleListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleViewResponse'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
      required: [items, page, size, totalItems]
    CreateScheduleRequest:
      type: object
      properties:
        scheduleKey:
          type: string
        selectedJobKey:
          type: string
        expression:
          type: string
        timezone:
          type: string
        enabled:
          oneOf:
            - type: boolean
            - type: 'null'
        description:
          type: string
      required: [scheduleKey, selectedJobKey, expression, timezone]
    UpdateScheduleRequest:
      type: object
      properties:
        selectedJobKey:
          type: string
        expression:
          type: string
        timezone:
          type: string
        enabled:
          oneOf:
            - type: boolean
            - type: 'null'
        description:
          type: string
      required: [selectedJobKey, expression, timezone]
    ScheduleStateChangeResponse:
      type: object
      properties:
        scheduleId:
          type: string
        enabled:
          type: boolean
        paused:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required: [scheduleId, enabled, paused, updatedAt]
...